+++ hata/discord/bases/entity/__init__.py
@@ -1,0 +1,10 @@
+from .discord_entity_meta import *
+from .entity_base import *
+from .slotted_meta import *
+
+
+__all__ = (
+    *discord_entity_meta.__all__,
+    *entity_base.__all__,
+    *slotted_meta.__all__,
+)